Command member variables set to (private) final
Like it is done in SubtractFlowCommand.
This revealed that the networkID in most command constructors is
actually never used. Instead, when calling execute(), the SNC's current
network is queried with snc.getCurrentNetwork().

An exception is EndToEndConvolutionCommand. It gets the current network
in its constructor and stores it in a member variable.

AddFlowCommand's success was never set to true. This is fixed.